Fix bug on macOS, remove .tmux.conf.func

+++ install.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 
+rm -rf ~/.mytmux
+
 # backup original tmux configuration
-mv $HOME/.tmux.conf $HOME/.tmux.conf.$(date "+%Y.%m.%d-%H:%M:%S")
-mv $HOME/.tmux.conf.func $HOME/.tmux.conf.func.$(date "+%Y.%m.%d-%H:%M:%S")
+if [ -f $HOME/.tmux.conf ]; then
+  mv $HOME/.tmux.conf $HOME/.tmux.conf.$(date "+%Y.%m.%d-%H:%M:%S")
+fi
 
 # clone from github
 git clone https://github.com/res876ttb/mytmux.git ~/.mytmux
 cd ~/.mytmux
 
 # creaet symbolic link to $HOME
-ln -s $PWD/.tmux.conf* $HOME/
-
+ln -s $PWD/.tmux.conf $HOME/
